backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
pymysql==1.1.1
alembic==1.13.3
python-multipart==0.0.17
orjson==3.10.7
cryptography==43.0.3
python-dotenv==1.0.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
